justeece/requirements.txt
boto3==1.9.157
django-storages==1.7.1
Django==2.2.22
django-tinymce==2.6.0
mysqlclient==2.0.3
cryptography==2.6.1
python-decouple==3.2
Pillow==8.3.2 
APScheduler==3.8.1